fix: fixed path error

+++ services/executor/utils/storage.py
@@ -2,13 +2,11 @@ import os
 import shutil
 from pathlib import Path
 
-
 ROOT_DIR = Path(__file__).resolve().parent.parent
 
-
 def create_tree(file_path):
-    os.makedirs(file_path.rsplit("/",1)[0], exist_ok=True)
-
+    split = file_path.rsplit(os.sep, 1)
+    os.makedirs(split[0], exist_ok=True)
 
 # use this method to avoid Invalid cross-device link error
 # https://stackoverflow.com/questions/42392600/oserror-errno-18-invalid-cross-device-link
@@ -16,7 +14,6 @@ def create_tree(file_path):
 def move(old_file, new_file):
     shutil.copy(old_file, new_file)
     os.remove(old_file)
-
 
 class Storage:
     def __init__(self, storage_path=None):
@@ -29,7 +26,11 @@ class Storage:
             self.path = ROOT_DIR_PROJECT.joinpath("storage")
 
     def abs_path(self, r_path):
-        return str(self.path.joinpath(r_path))
+        print("BLABLALBALA")
+        abs_p = os.path.join(str(self.path), r_path)
+        abs_p =f'/Executor/storage/{r_path}'
+        print(f"Resolved path: {abs_p}")
+        return str(abs_p)
 
     def move_to(self, l_file, s_file):
         s_abs_file = self.abs_path(s_file)
@@ -38,8 +39,14 @@ class Storage:
 
     def copy_from(self, s_file, l_file):
         s_abs_file = self.abs_path(s_file)
+        if not os.path.exists(s_abs_file):
+            print(f"Source file does not exist: {s_abs_file}")
+            # while True:
+            #     print(f"Trying to copy from {s_abs_file} to {l_file}")
+            return
         create_tree(l_file)
         shutil.copy(s_abs_file, l_file)
+        print(f"Copied from {s_abs_file} to {l_file}")
 
     def remove(self, s_file):
         os.remove(self.abs_path(s_file))
